requirements.txt
# Core
streamlit
torch
numpy
pandas
scikit-learn
plotly
vnstock

# Phase 1: Time Series & Statistics
statsmodels
hmmlearn
filterpy

# Phase 2: Network Analysis (coming)
networkx

# Phase 3: Multivariate (coming)
arch

# Utilities
scipy

# LlamaIndex for RAG Chatbot
llama-index>=0.12.0
llama-index-llms-openai
openai

# Configuration
python-dotenv
pydantic-settings
